+++ src/Dns/Dns/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Automatic DNS NameServer Delegation
+    - Create DNS zone cmdlet accepts parent zone name as additional optional parameter.
+    - Adds NS records in the parent zone for newly created child zone.
 
 ## Version 1.0.0
 * General availability of `Az.Dns` module